feat: local server button place
