+++ BrightID/src/components/Helpers/NodeModal.tsx
@@ -1,5 +1,6 @@
-import React from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import {
+  Alert,
   StyleSheet,
   Text,
   TouchableOpacity,
@@ -10,9 +11,13 @@ import { BlurView } from '@react-native-community/blur';
 import { useNavigation } from '@react-navigation/native';
 import { useTranslation } from 'react-i18next';
 import { isEqual } from 'lodash';
+import { create } from 'apisauce';
+import { NetworkInfo } from 'react-native-network-info';
+import Clipboard from '@react-native-community/clipboard';
+import httpBridge from 'react-native-http-bridge';
 import { useDispatch, useSelector } from '@/store/hooks';
-import { LIGHT_BLACK, ORANGE, WHITE, BLACK } from '@/theme/colors';
-import { DEVICE_LARGE } from '@/utils/deviceConstants';
+import { BLACK, GREEN, LIGHT_BLACK, ORANGE, WHITE } from '@/theme/colors';
+import { DEVICE_IOS, DEVICE_LARGE } from '@/utils/deviceConstants';
 import { fontSize } from '@/theme/fonts';
 import {
   clearBaseUrl,
@@ -23,6 +28,9 @@ import {
   selectDefaultNodeUrls,
 } from '@/reducer/settingsSlice';
 import { leaveAllChannels } from '@/components/PendingConnections/actions/channelThunks';
+import GraphQl from '@/components/Icons/GraphQl';
+import { getUserInfo } from '@/components/Onboarding/ImportFlow/thunks/channelUploadThunks';
+import { getExplorerCode } from '@/utils/explorer';
 
 const NodeModal = () => {
   const navigation = useNavigation();
@@ -66,6 +74,82 @@ const NodeModal = () => {
     );
   }
 
+  const [httpServerUrl, setHttpServerUrl] = useState('');
+
+  const password = useSelector((state) => state.user.password);
+
+  const toggleHttpServer = useCallback(async () => {
+    const port = 9025;
+    if (!httpServerUrl) {
+      httpBridge.start(port, 'http_service', async (request) => {
+        // you can use request.url, request.type and request.postData here
+        const headers = {
+          'Access-Control-Allow-Origin': '*',
+          'Access-Control-Allow-Headers': '*',
+          'Access-Control-Allow-Methods': '*',
+        };
+        const url = request.url.slice(request.url.indexOf('/'));
+        if (request.type === 'OPTIONS') {
+          httpBridge.respond(
+            request.requestId,
+            200,
+            'text/html; charset=utf-8',
+            undefined,
+            headers,
+          );
+        } else if (request.type === 'GET' && url === '/v1/info') {
+          httpBridge.respond(
+            request.requestId,
+            200,
+            'application/json',
+            JSON.stringify(await getUserInfo()),
+            headers,
+          );
+        } else if (request.type === 'GET' && url === '/v1/explorer-code') {
+          httpBridge.respond(
+            request.requestId,
+            200,
+            'application/json',
+            JSON.stringify({
+              explorerCode: getExplorerCode(),
+              password,
+            }),
+            headers,
+          );
+        } else {
+          httpBridge.respond(
+            request.requestId,
+            404,
+            'text/html; charset=utf-8',
+            '<html><body><a href="/v1/explorer-code" target="_blank">/v1/explorer-code</a></body></html>',
+            headers,
+          );
+        }
+      });
+      if (DEVICE_IOS) {
+        // to keep the server alive
+        create({
+          baseURL: `http://localhost:${port}`,
+        })
+          .get('/')
+          .catch(console.error);
+      }
+      const ip = await NetworkInfo.getIPV4Address();
+      const serverUrl = `${ip}:${port}`;
+      Clipboard.setString(serverUrl);
+      Alert.alert(t('home.alert.text.copied'));
+      setHttpServerUrl(serverUrl);
+    } else {
+      httpBridge.stop();
+      setHttpServerUrl('');
+    }
+  }, [httpServerUrl, password, t]);
+  useEffect(() => {
+    return () => {
+      httpBridge.stop();
+    };
+  }, []);
+
   return (
     <View style={styles.container}>
       <BlurView
@@ -94,6 +178,43 @@ const NodeModal = () => {
           </Text>
         </TouchableOpacity>
         {resetContainer}
+        <TouchableOpacity
+          testID="httpServerBtn"
+          style={[
+            styles.switchNodeButton,
+            styles.httpServerButton,
+            {
+              backgroundColor: httpServerUrl ? GREEN : ORANGE,
+            },
+          ]}
+          onPress={toggleHttpServer}
+          accessible={true}
+          accessibilityLabel={t('home.button.httpServer')}
+        >
+          <GraphQl
+            width={DEVICE_LARGE ? 25 : 20}
+            height={DEVICE_LARGE ? 25 : 20}
+          />
+          {httpServerUrl ? (
+            <View style={styles.httpServerInfo}>
+              <Text
+                style={[
+                  styles.switchNodeButtonText,
+                  {
+                    color: BLACK,
+                  },
+                ]}
+              >
+                {httpServerUrl}
+              </Text>
+              <Text style={styles.wifiSharingText}>WiFi Sharing Url</Text>
+            </View>
+          ) : (
+            <Text style={styles.switchNodeButtonText}>
+              {t('home.button.httpServer')}
+            </Text>
+          )}
+        </TouchableOpacity>
       </View>
     </View>
   );
@@ -147,6 +268,7 @@ const styles = StyleSheet.create({
     borderRadius: 20,
   },
   switchNodeButtonText: {
+    paddingLeft: 4,
     fontFamily: 'Poppins-Medium',
     fontSize: fontSize[15],
     color: WHITE,
@@ -173,6 +295,23 @@ const styles = StyleSheet.create({
     fontFamily: 'Poppins-Medium',
     fontSize: fontSize[14],
     color: WHITE,
+  },
+  httpServerButton: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    flexDirection: 'row',
+    marginTop: DEVICE_LARGE ? 8 : 6,
+  },
+  httpServerInfo: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    flexDirection: 'column',
+  },
+  wifiSharingText: {
+    fontFamily: 'Poppins-Medium',
+    fontSize: fontSize[10],
+    color: BLACK,
+    lineHeight: DEVICE_LARGE ? 12 : 10,
   },
 });
