+++ BrightID/src/components/HomeScreen.tsx
@@ -1,16 +1,15 @@
 import React, { useCallback, useContext, useEffect, useState } from 'react';
 import {
   ActivityIndicator,
-  Alert,
   Image,
   Linking,
-  StatusBar,
   StyleSheet,
   Text,
   TouchableOpacity,
   View,
+  StatusBar,
+  Alert,
 } from 'react-native';
-import { NetworkInfo } from 'react-native-network-info';
 import Clipboard from '@react-native-community/clipboard';
 import { createSelector } from '@reduxjs/toolkit';
 import { useFocusEffect } from '@react-navigation/native';
@@ -18,36 +17,34 @@ import { useHeaderHeight } from '@react-navigation/stack';
 import { useActionSheet } from '@expo/react-native-action-sheet';
 import { useTranslation } from 'react-i18next';
 import Material from 'react-native-vector-icons/MaterialCommunityIcons';
-import httpBridge from 'react-native-http-bridge';
-import { create } from 'apisauce';
 import { useDispatch, useSelector } from '@/store/hooks';
 import {
   fetchApps,
-  selectActiveDevices,
   setActiveNotification,
   updateBlindSigs,
+  selectActiveDevices,
 } from '@/actions';
 import { linkedContextTotal } from '@/reducer/appsSlice';
 import { verifiedConnectionsSelector } from '@/reducer/connectionsSlice';
 import { retrieveImage } from '@/utils/filesystem';
-import { BLACK, BLUE, DARKER_GREY, GREEN, ORANGE, WHITE } from '@/theme/colors';
+import { WHITE, ORANGE, BLACK, BLUE, DARKER_GREY } from '@/theme/colors';
 import fetchUserInfo from '@/actions/fetchUserInfo';
 import ChatBox from '@/components/Icons/ChatBox';
 import UnverifiedSticker from '@/components/Icons/UnverifiedSticker';
 import Camera from '@/components/Icons/Camera';
-import { DEVICE_IOS, DEVICE_LARGE } from '@/utils/deviceConstants';
+import { DEVICE_LARGE } from '@/utils/deviceConstants';
 import { fontSize } from '@/theme/fonts';
 import { setHeaderHeight } from '@/reducer/walkthroughSlice';
 import {
-  removeCurrentNodeUrl,
   selectBaseUrl,
   selectIsPrimaryDevice,
+  removeCurrentNodeUrl,
 } from '@/reducer/settingsSlice';
 import { NodeApiContext } from '@/components/NodeApiGate';
 import { getVerificationPatches } from '@/utils/verifications';
 import {
-  selectCompletedTaskIds,
   selectTaskIds,
+  selectCompletedTaskIds,
 } from '@/components/Tasks/TasksSlice';
 
 import { version as app_version } from '../../package.json';
@@ -56,9 +53,6 @@ import {
   syncAndLinkSocialMedias,
   updateSocialMediaVariations,
 } from '@/components/EditProfile/socialMediaThunks';
-import { getUserInfo } from '@/components/Onboarding/ImportFlow/thunks/channelUploadThunks';
-import GraphQl from '@/components/Icons/GraphQl';
-import { getExplorerCode } from '@/utils/explorer';
 
 /**
  * Home screen of BrightID
@@ -241,78 +235,6 @@ export const HomeScreen = (props) => {
     </View>
   ) : null;
 
-  const [httpServerUrl, setHttpServerUrl] = useState('');
-  const toggleHttpServer = useCallback(async () => {
-    const port = 9025;
-    if (!httpServerUrl) {
-      httpBridge.start(port, 'http_service', async (request) => {
-        // you can use request.url, request.type and request.postData here
-        const headers = {
-          'Access-Control-Allow-Origin': '*',
-          'Access-Control-Allow-Headers': '*',
-          'Access-Control-Allow-Methods': '*',
-        };
-        const url = request.url.slice(request.url.indexOf('/'));
-        if (request.type === 'OPTIONS') {
-          httpBridge.respond(
-            request.requestId,
-            200,
-            'text/html; charset=utf-8',
-            undefined,
-            headers,
-          );
-        } else if (request.type === 'GET' && url === '/v1/info') {
-          httpBridge.respond(
-            request.requestId,
-            200,
-            'application/json',
-            JSON.stringify(await getUserInfo()),
-            headers,
-          );
-        } else if (request.type === 'GET' && url === '/v1/explorer-code') {
-          httpBridge.respond(
-            request.requestId,
-            200,
-            'application/json',
-            getExplorerCode(),
-            headers,
-          );
-        } else {
-          httpBridge.respond(
-            request.requestId,
-            404,
-            'application/json',
-            '{"message": "not found"}',
-            headers,
-          );
-        }
-      });
-      if (DEVICE_IOS) {
-        // to keep the server alive
-        create({
-          baseURL: `http://localhost:${port}`,
-        })
-          .get('/')
-          .catch(console.error);
-      }
-      const ip = await NetworkInfo.getIPV4Address();
-      const serverUrl = `${ip}:${port}`;
-      Clipboard.setString(serverUrl);
-      Alert.alert(t('home.alert.text.copied'));
-      setHttpServerUrl(serverUrl);
-    } else {
-      httpBridge.stop();
-      setHttpServerUrl('');
-    }
-  }, [httpServerUrl]);
-  useEffect(() => {
-    return () => {
-      httpBridge.stop();
-    };
-  }, []);
-
-  const password = useSelector((state) => state.user.password);
-
   return (
     <View style={[styles.container, { marginTop: headerHeight }]}>
       <StatusBar
@@ -472,38 +394,6 @@ export const HomeScreen = (props) => {
               height={DEVICE_LARGE ? 25 : 20}
             />
             <Text style={styles.connectText}>{t('home.button.scanCode')}</Text>
-          </TouchableOpacity>
-          <TouchableOpacity
-            testID="httpServerBtn"
-            style={[
-              styles.connectButton,
-              styles.httpServerButton,
-              {
-                backgroundColor: httpServerUrl ? GREEN : WHITE,
-              },
-            ]}
-            onPress={toggleHttpServer}
-            accessible={true}
-            accessibilityLabel={t('home.button.httpServer')}
-          >
-            <GraphQl
-              width={DEVICE_LARGE ? 25 : 20}
-              height={DEVICE_LARGE ? 25 : 20}
-            />
-            {httpServerUrl ? (
-              <View style={styles.httpServerInfo}>
-                <Text style={styles.connectText}>
-                  {httpServerUrl || t('home.button.httpServer')}
-                </Text>
-                <Text style={styles.passwordText}>
-                  {password && `password: ${password}`}
-                </Text>
-              </View>
-            ) : (
-              <Text style={styles.connectText}>
-                {t('home.button.httpServer')}
-              </Text>
-            )}
           </TouchableOpacity>
           <TouchableOpacity
             testID="JoinCommunityBtn"
@@ -703,21 +593,6 @@ const styles = StyleSheet.create({
     shadowRadius: 10,
     elevation: 1,
     marginBottom: DEVICE_LARGE ? 16 : 11,
-  },
-  httpServerButton: {
-    marginTop: DEVICE_LARGE ? 8 : 6,
-  },
-  httpServerInfo: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    flexDirection: 'column',
-  },
-  passwordText: {
-    fontFamily: 'Poppins-Bold',
-    fontSize: fontSize[10],
-    color: BLACK,
-    marginLeft: DEVICE_LARGE ? 10 : 8,
-    lineHeight: DEVICE_LARGE ? 14 : 12,
   },
   connectText: {
     fontFamily: 'Poppins-Bold',
